requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
defusedxml==0.7.1
deprecated==1.2.18
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fonttools==4.58.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
oscrypto==1.3.0
packaging==25.0
pikepdf==9.7.0
pillow==11.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyhanko==0.28.0
pyhanko-certvalidator==0.27.0
pyjwt==2.10.1
python-barcode==0.15.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-pkcs11==0.7.0
pyyaml==6.0.2
qrcode==8.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typer==0.15.2
typing-extensions==4.12.2
typst==0.13.1
tzlocal==5.3.1
uharfbuzz==0.50.2
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
wrapt==1.17.2
